+++ src/rinoh/image.py
@@ -23,6 +23,7 @@ from .inline import InlineFlowable
 from .layout import ContainerOverflow, EndOfContainer
 from .number import NumberFormat
 from .paragraph import StaticParagraph, Paragraph, ParagraphStyle
+from .strings import StringCollection, String, StringField
 from .structure import ListOf, ListOfSection
 from .text import MixedStyledText, SingleStyledText, TextStyle, ErrorText
 from .util import posix_path, ReadAliasAttribute, PeekIterator
@@ -397,7 +398,8 @@ class Caption(StaticParagraph):
     def text(self, container):
         try:
             number = self.number(container)
-            label = [self.referenceable.category, ' ', number]
+            category_label = StringField(FloatLabels, self.referenceable.category)
+            label = [category_label, ' ', number]
         except KeyError:
             label = []
         return MixedStyledText(label + [self.content], parent=self)
@@ -409,12 +411,19 @@ class FigureStyle(FloatStyle, GroupedFlowablesStyle):
 
 class Figure(Float, StaticGroupedFlowables):
     style_class = FigureStyle
-    category = 'Figure'
+    category = 'figure'
 
 
 class ListOfFigures(ListOf):
-    category = 'Figure'
+    category = 'figure'
 
 
 class ListOfFiguresSection(ListOfSection):
     list_class = ListOfFigures
+
+
+class FloatLabels(StringCollection):
+    """Collection of localized titles for common sections"""
+
+    figure = String('Caption label for figures')
+    table = String('Caption label for tables')